Add functionality to record tab position in Waypoints.
Text editors that were newly created don't have a file name (or path), so
we can't distinguish them from other non-path tabs if we use this to
identify them. This commit adds a new member, tabIndex, to the Waypoint
data structure that allows us to track the tab that is actually being
referred to.

Refs #9.